+++ src/main/java/net/archers/ArchersMod.java
@@ -3,7 +3,7 @@ package net.archers;
 import net.archers.block.ArcherBlocks;
 import net.archers.config.ArchersItemConfig;
 import net.archers.config.Default;
-import net.archers.config.EnchantmentsConfig;
+import net.archers.config.TweaksConfig;
 import net.archers.effect.Effects;
 import net.archers.item.ArcherItems;
 import net.archers.item.Group;
@@ -16,6 +16,8 @@ import net.fabric_extras.structure_pool.api.StructurePoolConfig;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
+import net.minecraft.enchantment.Enchantments;
+import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
@@ -25,6 +27,7 @@ import net.spell_engine.api.item.trinket.SpellBooks;
 import net.spell_engine.api.loot.LootConfig;
 import net.spell_engine.api.loot.LootHelper;
 import net.spell_engine.api.spell.SpellContainer;
+import net.spell_power.api.enchantment.EnchantmentRestriction;
 import net.tinyconfig.ConfigManager;
 
 public class ArchersMod implements ModInitializer {
@@ -44,8 +47,8 @@ public class ArchersMod implements ModInitializer {
             .sanitize(true)
             .build();
 
-    public static ConfigManager<EnchantmentsConfig> tweaksConfig = new ConfigManager<>
-            ("tweaks", new EnchantmentsConfig())
+    public static ConfigManager<TweaksConfig> tweaksConfig = new ConfigManager<>
+            ("tweaks", new TweaksConfig())
             .builder()
             .setDirectory(ID)
             .sanitize(true)
@@ -68,6 +71,10 @@ public class ArchersMod implements ModInitializer {
         Effects.register();
         registerVillages();
         subscribeEvents();
+
+        if (tweaksConfig.value.enable_infinity_for_crossbows) {
+            EnchantmentRestriction.permit(Enchantments.INFINITY, itemStack -> itemStack.getItem() instanceof CrossbowItem);
+        }
     }
 
     private void registerItemGroup() {