+++ src/main/java/net/archers/config/TweaksConfig.java
@@ -1,7 +1,8 @@
 package net.archers.config;
 
-public class EnchantmentsConfig { public EnchantmentsConfig() {}
+public class TweaksConfig { public TweaksConfig() {}
     public float power_enchantment_multiplier_per_level = 0.08F;
     public float quick_charge_enchantment_multiplier_per_level = 0.1F;
     public float serverside_crossbow_charging_tolerance = 0.95F;
+    public boolean enable_infinity_for_crossbows = true;
 }