+++ src/main/java/net/archers/mixin/item/CrossbowItemMixin.java
@@ -95,7 +95,6 @@ public class CrossbowItemMixin implements CustomRangedWeaponProperties {
         }
     }
 
-
     /**
      * Apply custom velocity
      */