+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.6
 
+- Add feature: Allow Infinity enchantment for CrossBows (configurable in: `config/archers/tweaks.json`)
 - Add item tag: `archers:spears`
 - Update Italian translation, thanks to Zano1999 #10
 - Update Russian translation, thanks to skel39eek66 #9